+++ src/taoensso/timbre/profiling.clj
@@ -27,7 +27,8 @@
   of named body. Always returns the body's result."
   [id & body] `(pspy* ~id (fn [] ~@body)))
 
-(defmacro p [id & body] `(pspy ~id ~@body)) ; Alias
+(defmacro p  [id & body] `(pspy  ~id ~@body)) ; Alias
+(defmacro p* [id & body] `(pspy* ~id ~@body)) ; Alias
 
 (comment
   (time (dotimes [_ 1000000])) ; ~20ms